+++ lib/active_record_compose/transaction_support.rb
@@ -7,16 +7,22 @@ module ActiveRecordCompose
 
     module ClassMethods
       def lease_connection
-        if ActiveRecord::Base.respond_to?(:lease_connection)
-          __skip__ = ActiveRecord::Base.lease_connection
+        if ar_class.respond_to?(:lease_connection)
+          __skip__ = ar_class.lease_connection
         else
-          ActiveRecord::Base.connection
+          ar_class.connection
         end
       end
 
-      def connection = ActiveRecord::Base.connection
+      def connection = ar_class.connection
+
+      def with_connection(&) = __skip__ = ar_class.with_connection(&)
 
       def composite_primary_key? = false
+
+      private
+
+      def ar_class = ActiveRecord::Base
     end
 
     def id = nil